requirements.txt
asn1crypto==0.24.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
Django==2.1.1
djangorestframework==3.8.2
et-xmlfile==1.0.1
fpdf==1.7.2
#get==1.0.3
idna==2.7
image==1.5.27
img2pdf==0.3.3
jdcal==1.4
numpy==1.15.2
opencv-contrib-python==4.1.1.26
openpyxl==2.5.8
pandas==0.23.4
pdf2image==1.9.0
Pillow==6.1.0
#post==1.0.2
public==1.0.3
pycparser==2.19
PyMySQL==0.9.2
PyPDF2==1.26.0
python-barcode==0.10.0
python-dateutil==2.7.3
python-Levenshtein==0.12.0
pytz==2018.5
pyzbar==0.1.8
query-string==1.0.2
reportlab==3.5.23
#request==1.0.2
requests==2.19.1
six==1.11.0
urllib3==1.23
xlrd==1.2.0
XlsxWriter==1.1.1
celery==4.3.0
redis ==3.3.11
xhtml2pdf==0.2.4
boto3==1.13.15
simple-crypt
xmltodict==0.12.
Hashids==1.3.1
PyJWT==1.3.0
djangorestframework_simplejwt==4.4.0
environs==8.0.0
tabulate==0.8.7
num2words==0.5.10
apscheduler==2.1.2
camelcase==0.2